fix(command): stop argument parsing on non-provided optional argument
The command framework will now stop parsing command arguments when it
encounters an optional argument for which the command sender hasn't
provided a value. This is considered a fix due to being a preferred
behavior: it allows having required command arguments after an optional
one. If the optional argument is provided, the required arguments after
it will have to be provided by the command sender too.
An example use case for this is a command that optionally takes in a
location: the location can be represented with an optional argument for
the X coordinate, followed by 2 required arguments for the Y and Z
coordinates respectively, so that command senders will have to either
not provide any coordinates, or provide a complete location.